olimpiada/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
dj-database-url==2.2.0
Django==4.0
djangorestframework==3.13.1
gunicorn==23.0.0
packaging==24.1
psycopg2-binary==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2
whitenoise==6.7.0
